Update and rename functions.py to utils.py
